+++ test/errormessage.cpp
@@ -9,14 +9,14 @@ ErrorMessage::ErrorMessage(quint8 seq, ErrorCode errorCode, quint8 seqError) :
 }
 
 ErrorMessage::ErrorMessage(const QByteArray &rawData) :
-	AbstractMessage(rawData.size() > 4 ? rawData.at(2) : 0)
+	AbstractMessage(rawData.size() > 4 ? static_cast<quint8>(rawData.at(2)) : 0)
 {
 	if (rawData.size() == 5 || rawData.size() == 6)
 	{
 		m_data.append(rawData.mid(4, rawData.size() - 4));
 	}
 
-	if (rawData.at(0) != identifier() || rawData.at(rawData.size() - 1) != checksum())
+	if (rawData.at(0) != identifier() || static_cast<quint8>(rawData.at(rawData.size() - 1)) != checksum())
 	{
 		m_data.clear(); // invalidate message
 	}
