+++ test/versionmessage.cpp
@@ -8,16 +8,16 @@ VersionMessage::VersionMessage(quint8 seq) :
 }
 
 VersionMessage::VersionMessage(const QByteArray &rawData) :
-	AbstractMessage(rawData.size() == 12 ? rawData.at(2) : 0)
+	AbstractMessage(rawData.size() == 12 ? static_cast<quint8>(rawData.at(2)) : 0)
 {
 	if (rawData.size() == 12)
 	{
-		if (rawData.at(0) == identifier())
+		if (static_cast<quint8>(rawData.at(0)) == identifier())
 		{
 			m_data.append(rawData.mid(3, 8));
 		}
 	
-		if (checksum() != rawData.at(11))
+		if (checksum() != static_cast<quint8>(rawData.at(11)))
 		{
 			m_data.clear(); //invalidate message
 		}
